Update qaconsole page structure
